Added signature checker contract

+++ src/SignatureChecker.sol
@@ -6,15 +6,12 @@ contract SignatureChecker {
     mapping(address => bool) public validAddresses;
 
     constructor(address[] memory _validAddresses) {
-        for (uint i = 0; i < _validAddresses.length; i++) {
+        for (uint256 i = 0; i < _validAddresses.length; i++) {
             validAddresses[_validAddresses[i]] = true;
         }
     }
 
-    function checkSigner(
-        string memory message,
-        bytes memory signature
-    ) public view returns (bool) {
+    function checkSigner(string memory message, bytes memory signature) public view returns (bool) {
         // Calculate the Ethereum Signed Message hash
         bytes32 messageHash = getEthSignedMessageHash(message);
 
@@ -25,10 +22,7 @@ contract SignatureChecker {
         return validAddresses[signer];
     }
 
-    function recoverSigner(
-        bytes32 hash,
-        bytes memory signature
-    ) internal pure returns (address) {
+    function recoverSigner(bytes32 hash, bytes memory signature) internal pure returns (address) {
         // Ensure the signature length is valid
         require(signature.length == 65, "Invalid signature length");
 
@@ -50,16 +44,8 @@ contract SignatureChecker {
         return recoveredAddress;
     }
 
-    function getEthSignedMessageHash(
-        string memory message
-    ) public pure returns (bytes32) {
+    function getEthSignedMessageHash(string memory message) public pure returns (bytes32) {
         bytes32 messageHash = keccak256(abi.encodePacked(message));
-        return
-            keccak256(
-                abi.encodePacked(
-                    "\x19Ethereum Signed Message:\n32",
-                    messageHash
-                )
-            );
+        return keccak256(abi.encodePacked("\x19Ethereum Signed Message:\n32", messageHash));
     }
 }
